requirements.txt
anthropic
tavily-python
streamlit
python-dotenv
yfinance
pandas